Alertmanager: Fix computeConfig warning log (#12450)
We're currently logging `"merging configurations not implemented, using
mimir config"` whenever a tenant has a custom Mimir configuration and a
Grafana configuration that's not marked as default.

If the tenant doesn't have a Grafana config, the struct will have the
`Default` field set to `false` (zero value for booleans), and the line
will be logged. We can fix this by checking for promoted configs instead
of non-default configs.