index planning: implement cost model plan (#12199)
This is a follow-up of https://github.com/grafana/mimir/pull/12197

This PR introduces the plan. The plan is a collection of predicates.
Some of which are used when reading the index and the rest are used for
a sequential scan.


The code right now is not used, but will be when
https://github.com/grafana/mimir/issues/11919 is finished

### Cost

The plan has a cost model. The model consist of

- **index lookup cost**: performing index lookups for all predicates
that use the index
- **intersection cost**: cost of intersecting posting lists from
multiple index predicates
- **filter cost**: sequential scan i.e. cost of applying scan predicates
to the fetched series)


### Testing
The tests here use the same pattern as the ones in
https://github.com/grafana/mimir/pull/12197. In short:

1. Hard-coded mock statistics
3. CSV file instead of typical golang table tests.
   

#### Which issue(s) this PR fixes or relates to

related to https://github.com/grafana/mimir/issues/11919

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: chencs <casie.chen@grafana.com>