docs: ha-tracker allow usage of memberlist as kv store. (#12451)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

With version 2.17 we marked Memberlist as stable kv store for the Ha
tracker, and marked etcd,consul as deprecated.
We update the docs so it represents that

#### Which issue(s) this PR fixes or relates to

Part of
https://github.com/grafana/mimir/issues/1597#issuecomment-3200265688

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.