MQE: separate query planner for querier and query-frontends (#12449)
#### What this PR does

This PR introduces a separate query planner for queriers and another for
query-frontends.

Remote execution of query plans is implemented as an optimisation pass
that identifies subtrees of the query plan that should be executed
remotely. This optimisation pass must only run in query-frontends:
queriers should not remotely execute parts of the query plan themselves.

Therefore, in order to behave correctly when running in monolithic or
read/write mode, we need to have separate query planners for queriers
and query-frontends.

I have not added a changelog entry given this is a refactoring.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.