rw2: Do not silently drop metric metadata of type Unknown (#12461)
#### What this PR does

Fixes a bug where we silently drop a metric metadata, if the type is
`UNKNOWN`. Such metadata might still contain help/unit text that we want
to store. This affects both the Remote Write 2.0 push API and
ingest-storage V2 record format.

This happens because we normalize the metric name at the moment we parse
the type. Since `UNKNOWN` is the default metric type, Proto often
chooses to save a few bytes and never send the field at all. That means
we also never parse the `type` field, thus we never normalize the name,
thus the value of `normalizeMetricName` is always empty string. The
parser now thinks the metric name is empty and drops the metadata as it
maps to no series.

The fix simply normalizes the metric name a little later, outside of the
`metricType` parsing branch.

#### Which issue(s) this PR fixes or relates to

Contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.