Add CLI flag to set max Alertmanager notification batch size (#12469)
#### What this PR does

This PR adds a CLI flag to the ruler to allow configuring the max
Alertmanager notification batch size.

The default value (256) is the same as the value currently used, so
users that don't set this flag should see identical behaviour to before.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.