Usage tracker service (pre-kafka tenant active series limits) (#12358)
#### What this PR does

This PR creates a new service called usage-tracker.

The purpose of the usage tracker is to enforce active series limits
before ingesting the data into Kafka in the new ingest storage
architecture.

This service is basically a distributed, coordinated and replicated
in-memory hashmap storing series hashes for each one of the tenant.

This PR is huge, this has been a while of development, so I've come up
with some instructions for review:

- First of all, please read the doc in `docs/internal/usage-tracker.md`,
they describe how the service works, the expectations and the decisions
taken.
- After that, I would recommend proceeding with review from inside out:
- `pkg/usagetracker/tenantshard` package contains the implementation of
the in-memory swiss-like hashmap: this is an implementation of
`map[hash]timestamp` optimized for our needs:
- Our data is small, we store timestamps as "minutes since the last 2h
multiple"
- We don't care about values higher than 1h anyway: that's the maximum
active series TTL this service allows to be set. We also don't care
about higher precision: we do the cleanup once in a minute, so
sub-minute precision is useless here.
- The code is optimized to perform the update logic in the map itself,
so we don't have to retrieve, compare, and put back the value for each
series hash.
- This is almost everything this service does, so it's worth optimizing
it.
- The code can peform garbage collection by just iterating the values
(not the keys), exploiting cache locality even more.
- `pkg/usagetracker/clock` just contains a type-wrapper for our
minutes-timestamp.
- `pkg/usagetracker/tracker_store.go` is the core business logic of the
usage-tracker. It handles the track series requests, creating the
tenantshard instances, and also processes the incoming "created series"
events.
- `pkg/usagetracker/tracker_store_snapshot.go` extends that with methods
necessary to genearte & load snapshots.
- `pkg/usagetracker/tracker_store_collector.go` implements a prometheus
collector, so each tracker store can be registered in the prometheus
registry directly.
- `pkg/usagetracker/partition_handler.go` is the `service.Service`
handling the lifecycle of one partition in the usage-tracker service
replica. It holds one tracker store, and adds layers of event
processing, snapshot coordination, etc.
- `pkg/usagetracker/tracker.go` is the main usage-tracker service file:
it coordinates the creation and removal of partition handlers, as well
as the general instance read-only state and shutdown process.
- Finally, `pkg/usagetracker/tracer_instance_ring.go` and
`pkg/usagetracker/tracer_partition_ring.go` are the usual boilerplate
for both instance and partition ring used by usage-tracker.

Finally, for the distributor/client part, check out the code in
`pkg/usagetracker/usagetrackerclient` and
`prePushMaxSeriesLimitMiddleware` in the distributor.

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>
Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>
Co-authored-by: Toni Cárdenas <toni@tcardenas.me>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>