helm: embed demo kafka cluster for ingest storage (#12459)
#### What this PR does

The PR adds a demo Kafka cluster (disable by default, **for now**), that
will be used when the chart will deploy Mimir in the ingest storage
architecture.

This is the first step for what we've agreed in
https://github.com/grafana/mimir/issues/12177

> We will embed a demo Kafka cluster (enabled by default), with the same
"don't use it in production" notice as Minio

The PR also adds a new test file `test-ingest-storage-values.yaml`, that
configures Mimir to run in the ingest storage mode, and use the embedded
Kafka.

I've tested it in a local k8s cluster:

```
helm upgrade mimir . --install --create-namespace --namespace test-mimir -f ci/offline/test-ingest-storage-values.yaml
```

The above created a new Mimir cluster, and connected the write- and
read-components through the embedded single-node Kafka cluster. Both
read and writes to the Mimir cluster worked as I expected.

#### Which issue(s) this PR fixes or relates to

Relates to https://github.com/grafana/mimir/issues/12177
https://github.com/grafana/mimir/issues/4560

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>