Multiple Components: Updates to CPU and Memory HPA query (#12406)
#### What this PR does

At Grafana Labs, we have encountered cases where rate() extrapolates a
metric over missing values, underestimates CPU utilization, and triggers
an unnecessary downscaling. This PR switches the default CPU scaler
metrics to use `irate()` rather than `rate()`
(see:autoscaling.libsonnet) and affects the scaling behavior of the
following components:

- alertmanager
- query frontend
- ruler
- ruler querier
- ruler query frontend
- distributor

Because `irate()` can be more volatile, using the maximum of irate can
cause _over-scaling_. We mitigate this by using the p75 of the `irate()`
distribution. Mimir operators can opt-in to using `irate()` by setting a
new config option - `autoscaling_cpu_hpa_use_irate`.

We also remove `or vector(0)` from a leg of the memory metric. In the
first leg of the query, we remove `or vector(0)` so that if metrics are
unavailable, we part returns no data instead of zeroes to prevents
underscaling when metrics are unavailable. We keep `or vector(0)` on the
second leg because this part should always contribute a value to the
calculation. Mimir operators can opt-in to this change by setting a new
config option `autoscaling_memory_hpa_require_metrics`.

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.