Skip read path load shedding when an ingester is the only available replica (#12448)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

This change propagates some gRPC meta from queriers to ingesters
indicating when the ingester is the only replica being called. A server
middleware passes along this info in a context, which the ingester can
use to skip the read path's reactive limiter and circuit breaker. The
rationale for this is that even though an ingester may be overloaded,
when there is no other replica to go to, we may as well accept the load
in case it can be handled by overcommitting on resources.

Note: this is only applied when ingest storage is enabled, since that's
where reactive limiters are used.

#### Which issue(s) this PR fixes or relates to

Fixes #<issue number>

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.