MQE: simplify `RangeVectorOperator` interface (#12489)
#### What this PR does

This PR removes the `StepCount` and `Range` methods from the
`RangeVectorOperator` interface.

`StepCount` wasn't used anywhere, and `Range` was only used for a select
few range vector functions.

Despite now needing to compute the range on every step, removing `Range`
seems to have had a neutral (or possibly even slightly positive) impact
on latency in our benchmarks for range vector functions like `rate`.

I've opted not to add a changelog entry given this is an internal
refactoring with no user-visible impact.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.