Fix compactor_block_uploads_failed_total metric showing type="unknown" (#12477)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

This PR makes a small change to error handling in the block upload
logic, updates all returns of `UploadError` in `Upload()` to use
pointers (`&UploadError{...}`) instead of values.

When values are used,
[errors.As()](https://github.com/grafana/mimir/blob/4dde21906feff9da2646bcbc22fa2572d847c150/pkg/compactor/bucket_compactor.go#L474)
will not interpret the returned error as a `UploadError`, not updating
the metric with the specific label values and always adding the
label/value `type="unknown"`.


In addition, modifies the error wrapping in the `cleanUp` function to
use `fmt.Errorf` with `%w` for improved error propagation, replacing the
previous use of `errors.Wrapf`.

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Arve Knudsen <arve.knudsen@gmail.com>