MQE: introduce `Finalize` method to operators (#12492)
#### What this PR does

This PR introduces a new method to the `Operator` interface: `Finalize`.

`Finalize` is expected to be used to complete any outstanding work such
as emitting annotations or updating query stats before the result is
returned to the caller. The particular use case that inspired this is
remote execution: we need to apply any annotations and query stats from
the remote executor to the local query.

It differs from `Close` in that it is valid to call `Close` at any time,
even if the query has failed or if the result has already been sent to
the caller. In contrast, `Finalize` will only be called if the query has
succeeded, and before results are sent to the caller.

I haven't added a changelog entry given this is an internal refactoring.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Jon Kartago Lamida <jon.lamida@grafana.com>