MQE: Move DeduplicateAndMerge to planning stage  (#12470)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
This is the first batch of work for the Skipping DeduplicateAndMerge
operator when possible.
The PR introduces the DeduplicateAndMerge node and inserts it into the
planning stage so it can later be pruned with the optimization pass. The
optimization itself will be implemented in a next PR.

#### Implementation details

- A new DeduplicationAndMerge node type has been introduced. It simply
wraps the inner node -
[code](https://github.com/grafana/mimir/pull/12470/files#diff-a7515378fd6575121f3f2dd05dbdd8d0ad395afa1e8b19d9f8482dc5ba868670)
- During the planning stage, functions that drop name label, unary
negations, binary operations, and vector-scalar operations are wrapped
in DeduplicationAndMerge – -
[code](https://github.com/grafana/mimir/pull/12470/files#diff-8251d65134441870ca8a04b92479b72d640d81e53f86844fd87f18eb3aef5a13)
- Previously, the DeduplicationAndMerge operator was inserted in various
places (factories, operator constructors, etc.). This has been cleaned
up, and the operator is now created when the plan is materialized.

#### Checklist

- [x] Tests updated.
- [ ] ~Documentation added~.
- [ ] ~`CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.~
- [ ]
~[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.~

---------

Signed-off-by: Innokentii Konstantinov <innokenty.konstantinov@grafana.com>
Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>