helm: replace kubectl image with alpine equivalent (#12498)
#### What this PR does

Starting 28/08/2025 Bitnami deprecates their free images. In order to
mitigate any interruptions we update the default values, to use
alternative Docker images.

In this PR we update the Docker image with kubectl to `alpine/kubectl`.
This image comes with a shell, which is required by the provisioning
job, that uses the image (i.e. that is why we don't use
`registry.k8s.io/kubectl`).

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>