ingest-storage: Disable normalization of metric names in metadata in record format V2 (#12499)
#### What this PR does

The RW2 deserialization contains some app logic - it mutates metric
names in metadata for certain types. For example, given a series that's
part of a classic histogram called `my_histogram_seconds_bucket`, the
unmarshaller will drop the `_bucket` suffix. This also means it dedupes
metadata for corresponding `_count` and `_sum` series.

For ingest storage, it's more predictable to have the record format be
as transparent as possible and not edit the data in unexpected ways.
That way, a RW1 request is still guaranteed to retain its original
behavior regardless of record version.

This PR adds an internal option to skip the normalization, and enables
it for the ingest storage record path.

I don't think the deduping is necessarily bad in general; removes
elements of the request, and all members of a histogram tend to have the
same metadata. But, if we decide to normalize for V1, we should do it in
the distributor, prior to sending the data through warpstream, not in
the ingester unmarshal. That way it's not implicit, and it has a chance
of shrinking the record size.

Requests sent from the RW2 ingestion API will already be deduped in the
distributor and this does not affect them.
No changelog until we mark the V2 format as stable, as usual.

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.