MQE: add context parameter to `RangeVectorOperator.NextStepSamples` (#12513)
#### What this PR does

This PR adds a `context.Context` parameter to
`RangeVectorOperator.NextStepSamples()`.

This will be used by the remote range vector execution operator, as
`NextStepSamples` may block waiting for the network, and so the context
will allow it to abort if the query is cancelled or times out.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.