docs: remove documentation for using query-frontend with Prometheus (#12517)
#### What this PR does

In Mimir 3 we've removed support for using the query-frontend in front
of a Prometheus compatible server. This mode of running the query-
frontend was rarely used or tested and so has been removed to simplify
the query-frontend. This change removes the documentation for the
feature.

#### Which issue(s) this PR fixes or relates to

Related #12191

#### Checklist

- [ ] Tests updated.
- [X] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>