Abandon failed jobs in block-builder worker/client. (#12522)
There are circumstances that may cause `BlockBuilder.consumeJob` to
return an error. We need to fail the job when that happens, otherwise
the block builder will keep the job among its active jobs until the
container exits, which will keep the scheduler from expiring the lease
and reassigning it to another worker.