Pull statistics into planner (#12530)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does
This PR pulls the periodically-updated head statistics introduced in
https://github.com/grafana/mimir-prometheus/pull/962 into the ingester.
This is passed to a `CostBasedPlanner`, which is now used instead of the
`ScanEmptyMatchersLookupPlanner` if
`-blocks-storage.tsdb.index-planning-lookup-enabled` is enabled.

#### Which issue(s) this PR fixes or relates to

Relates to https://github.com/grafana/mimir/issues/11919

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.