Refactor query-frontend round-tripper creation (#12529)
#### What this PR does

This PR refactors how the query-frontend round-tripper is created.

This is to allow substituting in an alternative to
`httpQueryRequestRoundTripperHandler` that performs query planning and
evaluates the query with MQE. (This will be added in a subsequent PR.)

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.