MQE: fix panic when evaluating a nested subquery where the parent has no steps (#12524)
#### What this PR does

This PR fixes an issue where evaluating a nested subquery can panic if
the parent subquery has no steps.

For example, in the query
`avg_over_time((count_over_time(foo[5m:2m]))[7m:1h])`, the parent is
looking for steps on the hour over the last 7m. But if the query is
being evaluated at a timestamp where it's at least 7 minutes past the
hour, then the last 7m doesn't include any steps and this can result in
a panic.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.