mimirpb: Preserve inconsistent metric metadata in RW1 to 2 conversion (#12541)
#### What this PR does

Occasionally RW1.0 requests will contain inconsistent metadata. That is,
they might contain multiple metadatas for the same series that don't
agree with each other, like a new and old description simultaneously.
The prometheus behavior when accepting RW1.0 is to not choose a winner;
both metadatas are stored and will appear in the metadata API until they
both age out.

RW2.0 tends to prevent this by attaching metadata to series. However,
for requests that were converted from 1.0 to 2.0 transparently, we want
to preserve the original behavior and ensure both metadata make it
through the system.

This PR adjusts the RW1.0 to 2.0 conversion to only deduplicate metadata
whose fields are exactly equal during the timeseries matching process.
The previous implementation would discard inconsistent metadata beyond
the first seen one. If we see multiple metadata that exactly match, we
save space and deduplicate it; if we see a metadata that is
inconsistent, we fabricate an extra timeseries and send both versions.

No changelog as it only affects record format V2, which is not marked
stable.

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.