Introduce `FromSamplesToFPoints` and `FromHistogramsToHPoints` (#12552)
#### What this PR does

This PR introduces two new methods to the `mimirpb` package, to convert
from the `mimirpb` types to their `promql` equivalents.

This has been extracted from a larger PR to try to make that PR smaller
and easier to review.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.