ruler: add support for remote rule evaluation over HTTP (#11415)
We're working on a multi-region Mimir deployment. Our approach is going
to be to run two separate Mimir clusters, with a proxy (testing with
[promxy](https://github.com/jacksontj/promxy) right now) to provide an
unified view.
 
This is working work OK in our initial tests, one of the missing pieces
is the ability for `ruler` to query the proxy. This PR adds the option
to do this.

#### Checklist

- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.