MQE: return `APIError` instances for non-execution errors (#12578)
#### What this PR does

This PR modifies MQE to return `APIError` instances when appropriate,
and modifies the querier's `Dispatcher` to translate these to the
appropriate Protobuf constant for transmission back to query-frontends
when remote execution is being used.

This ensures that the correct HTTP status code is used when returning an
error response for a query request evaluated with remote execution.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.